Simple query added to hook_form_alter to get the vocabulary title when editing classes or properties. Added the submit callback to form save to redirect directly to the vocabulary page.
"Hook_nodeapi"
Added variable  previous_class/property_title to hold the old title when a node is going to be edited.
Added the pre-process to get all the authors for a vocabulary node when this one going to be viewed.
Added a variable to the node object to control when a node it forced to update.
Fixed the bug regarding the property subproperties y the term array, it should be subpropertyof.

"_neologism_on_deleted_class_node"
Change the option to recibe the default superclass when remove a superclass to its superclass.

if( count($nclass->field_superclass2) == 1 ) {
	$value['evoc_term'] = DEFAULT_RDFS_SUPERCLASS;
}

To

if( count($nclass->field_superclass2) > 0 ) {
	$value['evoc_term'] = $node-field_superclass2[0]'evoc_term'];
}

Added the hook_removed_external_vocabulary($prefix) to do some action after a namespace/vocabulary has been removed from the evoc system.

function neologism_removed_external_vocabulary($prefix) {
}