SparqlEngine.php protected function joinResults($patternlist){} line 634
$patternlist[$pattern['unionWith']]['patternResult'] not initialized.

array_push() fail showing a warning like: array_push() [function.array-push]: First argument should be an array. 

So we need to check when $patternlist[$pattern['unionWith']]['patternResult'] is not initialized just asking if it is not an array. If it doesn’t, we need to initialize it, otherwise the value will not be added using array_push().